what the fuck
I swear, I give BuckleScript, ReasonML, Rescript, whatever the fuck
you wanna call it a try and it's SO DAMN RESISTENT to making
anything easy for me jesus christ. You can't name a fucking field
`type`?! Sure, there might be a way for me to do that but if I want
to find any useful documentation, I have to look up documentation
for ReasonML. Christ guys, I'm *trying* to learn your language
because I think it's a good idea but hot damn is the execution so
shit. I can't waste my time anymore with this nonsense. This is
such bull.